Removing more warnings.
git-svn-id: http://svn.seqan.de/seqan/trunk@9514 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/core/include/seqan/find/find_pex.h
@@ -434,7 +434,7 @@ SEQAN_CHECKPOINT
 
     // adjust start and end if they point over the edges of host(finder)
     s = (s < 0 ? 0 : s);
-    e = (e > static_cast<int>(length(host(finder))) ? length(host(finder)) : e);
+    e = (e > static_cast<int>(length(host(finder))) ? static_cast<int>(length(host(finder))) : e);
 
     THostSegment i(infix(host(mf),s,e));
     THSFinder f(i);
